Added a new test project for groups
Added a new test project for groups
