+++ test/DynamoCoreTests/DynamoCoreTests.csproj
@@ -68,6 +68,7 @@
   <ItemGroup>
     <Compile Include="AstBuilderTest.cs" />
     <Compile Include="ConverterTests.cs" />
+    <Compile Include="GroupTests.cs" />
     <Compile Include="CrashProtectionTests.cs" />
     <Compile Include="DynamoModelTestBase.cs" />
     <Compile Include="MessageLogTests.cs" />
